Starting Foundation navagation
I'm starting to implement the large, medium, small navagation.

The trick is to just generate the navagation list twice, once for the large and medium grid sizes, and once for the small grid size. Then hide one when the other is active. 

Next: Generate the small menu. There is only a placeholder there now.